requirements.txt
# Rebar System - Python Flask Backend
# V12.0 企业级架构版

# ========================================
# Web Framework
# ========================================
flask>=2.3.0
flask-cors>=4.0.0
flask-sqlalchemy>=3.1.0
flask-migrate>=4.0.0

# ========================================
# Database
# ========================================
pymysql>=1.1.0
cryptography>=41.0.0

# ========================================
# Image Processing
# ========================================
opencv-python>=4.8.0
pillow>=10.0.0
# 固定 numpy 版本以解决 pytesseract 兼容性问题
numpy==1.26.4

# ========================================
# OCR (文字识别)
# ========================================
pytesseract>=0.3.10

# ========================================
# Excel Export
# ========================================
openpyxl>=3.1.0
pandas>=2.0.0

# ========================================
# HTTP Requests & MinIO
# ========================================
requests>=2.31.0
minio>=7.2.0

# ========================================
# VLM 大模型 (CAD 图纸解析)
# ========================================
zhipuai>=2.0.0

# ========================================
# Utilities
# ========================================
python-dotenv>=1.0.0
werkzeug>=2.3.0
